calling_test/requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
click==8.2.1
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepgram-sdk==4.1.1
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.17.0
exceptiongroup==1.3.0
fastapi==0.116.1
flatbuffers==25.2.10
frozenlist==1.7.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
jiter==0.10.0
llvmlite==0.44.0
loguru==0.7.3
markdown==3.8.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy-extensions==1.1.0
numba==0.61.2
numpy==2.2.6
onnxruntime==1.20.1
openai==1.74.1
packaging==25.0
pillow==11.1.0
pipecat-ai==0.0.76
propcache==0.3.2
protobuf==5.29.5
pydantic==2.10.6
pydantic-core==2.27.2
pyloudnorm==0.1.1
python-dotenv==1.1.1
resampy==0.4.3
scipy==1.15.3
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.47.2
sympy==1.14.0
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspect==0.9.0
uvicorn==0.35.0
websockets==13.1
yarl==1.20.1
twilio>=9.0.0
azure-ai-contentsafety>=1.0.0
azure-identity>=1.14.0
azure-cognitiveservices-speech
azure-cosmos>=4.7.0
